+++ confidential-identities/build.gradle
@@ -1,7 +1,5 @@
-// Experimental Confidential Identities support for 1.0
-// This contains the prototype SwapIdentitiesFlow and SwapIdentitiesHandler, which can be used
-// for exchanging confidential identities as part of a flow, until a permanent solution is prepared.
-// Expect this module to be removed and merged into core in a later release.
+// This contains the SwapIdentitiesFlow which can be used for exchanging confidential identities as part of a flow.
+// TODO: Merge this into core: the original plan was to develop it independently but in practice it's too widely used to break compatibility now, as finance uses it.
 apply plugin: 'kotlin'
 apply plugin: CanonicalizerPlugin
 apply plugin: 'net.corda.plugins.publish-utils'
@@ -25,17 +23,6 @@ dependencies {
 
     // AssertJ: for fluent assertions for testing
     testCompile "org.assertj:assertj-core:$assertj_version"
-}
-
-cordapp {
-    targetPlatformVersion corda_platform_version.toInteger()
-    minimumPlatformVersion 1
-    workflow {
-        name "Corda Experimental Confidential Identities"
-        versionId 1
-        vendor "R3"
-        licence "Open Source (Apache 2)"
-    }
 }
 
 jar {
