+++ finance/workflows/build.gradle
@@ -28,8 +28,8 @@ dependencies {
     // cordapp project(':finance:workflows')
     // cordapp project(':finance:contracts')
     cordaCompile project(':core')
+    cordaCompile project(':confidential-identities')
     
-    cordapp project(':confidential-identities')
     cordapp project(':finance:contracts')
 
     // For JSON
